+++ tests/QA/module.tests.ps1
@@ -1,112 +1,207 @@
-$here = Split-Path -Parent $MyInvocation.MyCommand.Path
+BeforeDiscovery {
+    $script:moduleName = 'Sampler'
 
-# Convert-path required for PS7 or Join-Path fails
-$ProjectPath = "$here\..\.." | Convert-Path
-$ProjectName = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
-    ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) -and
-    $(try { Test-ModuleManifest $_.FullName -ErrorAction Stop }catch{$false}) }
-).BaseName
+    Remove-Module -Name $script:moduleName -Force -ErrorAction 'SilentlyContinue'
 
-$SourcePath = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
-        ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) -and
-        $(try { Test-ModuleManifest $_.FullName -ErrorAction Stop }catch { $false }) }
+    $mut = Get-Module -Name $script:moduleName -ListAvailable |
+        Select-Object -First 1 |
+        Import-Module -Force -ErrorAction 'Stop' -PassThru
+}
+
+BeforeAll {
+    $script:moduleName = 'Sampler'
+
+    # Convert-Path required for PS7 or Join-Path fails
+    $projectPath = "$($PSScriptRoot)\..\.." | Convert-Path
+
+    $sourcePath = (
+        Get-ChildItem -Path $projectPath\*\*.psd1 |
+            Where-Object -FilterScript {
+                ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) `
+                -and $(
+                    try
+                    {
+                        Test-ModuleManifest -Path $_.FullName -ErrorAction 'Stop'
+                    }
+                    catch
+                    {
+                        $false
+                    }
+                )
+            }
     ).Directory.FullName
-
-$mut = Import-Module -Name $ProjectName -ErrorAction Stop -PassThru -Force
-$allModuleFunctions = &$mut {Get-Command -Module $args[0] -CommandType Function } $ProjectName
+}
 
     Describe 'Changelog Management' -Tag 'Changelog' {
         It 'Changelog has been updated' -skip:(
             !([bool](Get-Command git -EA SilentlyContinue) -and
               [bool](&(Get-Process -id $PID).Path -NoProfile -Command 'git rev-parse --is-inside-work-tree 2>$null'))
-            ) {
-            # Get the list of changed files compared with main
+        ) {
+            # Get the list of changed files compared with branch main
             $HeadCommit = &git rev-parse HEAD
             $defaultBranchCommit = &git rev-parse origin/main
             $filesChanged = &git @('diff', "$defaultBranchCommit...$HeadCommit", '--name-only')
+            $filesStagedAndUnstaged = &git @('diff', "HEAD", '--name-only')
 
-            if ($HeadCommit -ne $defaultBranchCommit) { # if we're not testing same commit (i.e. main..main)
-                $filesChanged.Where{ (Split-Path $_ -Leaf) -match '^changelog' } | Should -Not -BeNullOrEmpty
-            }
+            $filesChanged += $filesStagedAndUnstaged
+
+            $filesChanged | Should -Contain 'CHANGELOG.md' -Because 'the CHANGELOG.md must be updated with at least one entry in the Unreleased section for each PR'
         }
 
         It 'Changelog format compliant with keepachangelog format' -skip:(![bool](Get-Command git -EA SilentlyContinue)) {
             { Get-ChangelogData (Join-Path $ProjectPath 'CHANGELOG.md') -ErrorAction Stop } | Should -Not -Throw
         }
-    }
 
-    Describe 'General module control' -Tags 'FunctionalQuality' {
-
-        It 'imports without errors' {
-            { Import-Module -Name $ProjectName -Force -ErrorAction Stop } | Should -Not -Throw
-            Get-Module $ProjectName | Should -Not -BeNullOrEmpty
-        }
-
-        It 'Removes without error' {
-            { Remove-Module -Name $ProjectName -ErrorAction Stop } | Should -not -Throw
-            Get-Module $ProjectName | Should -beNullOrEmpty
+        It 'Changelog should have an Unreleased header' -Skip:$skipTest {
+            (Get-ChangelogData -Path (Join-Path -Path $ProjectPath -ChildPath 'CHANGELOG.md') -ErrorAction 'Stop').Unreleased.RawData | Should -Not -BeNullOrEmpty
         }
     }
 
-    if (Get-Command Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue) {
-        $scriptAnalyzerRules = Get-ScriptAnalyzerRule
-    }
-    else {
-        if ($ErrorActionPreference -ne 'Stop') {
-            Write-Warning "ScriptAnalyzer not found!"
-        }
-        else {
-            Throw "ScriptAnalyzer not found!"
-        }
+Describe 'General module control' -Tags 'FunctionalQuality' {
+    It 'Should import without errors' {
+        { Import-Module -Name $script:moduleName -Force -ErrorAction Stop } | Should -Not -Throw
+
+        Get-Module -Name $script:moduleName | Should -Not -BeNullOrEmpty
     }
 
-    foreach ($function in $allModuleFunctions) {
-        $functionFile = Get-ChildItem -path $SourcePath -Recurse -Include "$($function.Name).ps1"
-        Describe "Quality for $($function.Name)" -Tags 'TestQuality' {
-            It "$($function.Name) has a unit test" {
-                Get-ChildItem "tests\" -recurse -include "$($function.Name).Tests.ps1" | Should Not BeNullOrEmpty
-            }
+    It 'Should remove without error' {
+        { Remove-Module -Name $script:moduleName -ErrorAction Stop } | Should -Not -Throw
 
-            if ($scriptAnalyzerRules) {
-                It "Script Analyzer for $($functionFile.FullName)" {
-                    $PSSAResult = (Invoke-ScriptAnalyzer -Path $functionFile.FullName)
-                    $Report = $PSSAResult | Format-Table -AutoSize | Out-String -Width 110
-                    $PSSAResult  | Should -BeNullOrEmpty -Because `
-                        "some rule triggered.`r`n`r`n $Report"
-                }
+        Get-Module $script:moduleName | Should -BeNullOrEmpty
+    }
+}
 
-            }
+BeforeDiscovery {
+    # Must use the imported module to build test cases.
+    $allModuleFunctions = & $mut { Get-Command -Module $args[0] -CommandType Function } $script:moduleName
+
+    # Build test cases.
+    $testCases = @()
+
+    foreach ($function in $allModuleFunctions)
+    {
+        $testCases += @{
+            Name = $function.Name
         }
+    }
+}
 
-        Describe "Help for $($function.Name)" -Tags 'helpQuality' {
-            $AbstractSyntaxTree = [System.Management.Automation.Language.Parser]::
-            ParseInput((Get-Content -raw $functionFile.FullName), [ref]$null, [ref]$null)
-            $AstSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
-            $ParsedFunction = $AbstractSyntaxTree.FindAll( $AstSearchDelegate, $true ) |
-                ? Name -eq $function.Name
-
-            $FunctionHelp = $ParsedFunction.GetHelpContent()
-
-            It 'Has a SYNOPSIS' {
-                $FunctionHelp.Synopsis | should not BeNullOrEmpty
+Describe 'Quality for module' -Tags 'TestQuality' {
+    BeforeDiscovery {
+        if (Get-Command -Name Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue)
+        {
+            $scriptAnalyzerRules = Get-ScriptAnalyzerRule
+        }
+        else
+        {
+            if ($ErrorActionPreference -ne 'Stop')
+            {
+                Write-Warning -Message 'ScriptAnalyzer not found!'
             }
-
-            It 'Has a Description, with length > 40' {
-                $FunctionHelp.Description.Length | Should beGreaterThan 40
-            }
-
-            It 'Has at least 1 example' {
-                $FunctionHelp.Examples.Count | Should beGreaterThan 0
-                $FunctionHelp.Examples[0] | Should match ([regex]::Escape($function.Name))
-                $FunctionHelp.Examples[0].Length | Should BeGreaterThan ($function.Name.Length + 10)
-            }
-
-            $parameters = $ParsedFunction.Body.ParamBlock.Parameters.name.VariablePath.Foreach{ $_.ToString() }
-            foreach ($parameter in $parameters) {
-                It "Has help for Parameter: $parameter" {
-                    $FunctionHelp.Parameters.($parameter.ToUpper()) | Should Not BeNullOrEmpty
-                    $FunctionHelp.Parameters.($parameter.ToUpper()).Length | Should BeGreaterThan 25
-                }
+            else
+            {
+                throw 'ScriptAnalyzer not found!'
             }
         }
     }
+
+    It 'Should have a unit test for <Name>' -TestCases $testCases {
+        Get-ChildItem -Path 'tests\' -Recurse -Include "$Name.Tests.ps1" | Should -Not -BeNullOrEmpty
+    }
+
+    It 'Should pass Script Analyzer for <Name>' -TestCases $testCases -Skip:(-not $scriptAnalyzerRules) {
+        $functionFile = Get-ChildItem -Path $sourcePath -Recurse -Include "$Name.ps1"
+
+        $pssaResult = (Invoke-ScriptAnalyzer -Path $functionFile.FullName)
+        $report = $pssaResult | Format-Table -AutoSize | Out-String -Width 110
+        $pssaResult  | Should -BeNullOrEmpty -Because `
+            "some rule triggered.`r`n`r`n $report"
+    }
+}
+
+Describe 'Help for module' -Tags 'helpQuality' {
+    It 'Should have .SYNOPSIS for <Name>' -TestCases $testCases {
+        $functionFile = Get-ChildItem -Path $sourcePath -Recurse -Include "$Name.ps1"
+
+        $scriptFileRawContent = Get-Content -Raw -Path $functionFile.FullName
+
+        $abstractSyntaxTree = [System.Management.Automation.Language.Parser]::ParseInput($scriptFileRawContent, [ref] $null, [ref] $null)
+
+        $astSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
+
+        $parsedFunction = $abstractSyntaxTree.FindAll( $astSearchDelegate, $true ) |
+            Where-Object -FilterScript {
+                $_.Name -eq $Name
+            }
+
+        $functionHelp = $parsedFunction.GetHelpContent()
+
+        $functionHelp.Synopsis | Should -Not -BeNullOrEmpty
+    }
+
+    It 'Should have a .DESCRIPTION with length greater than 40 characters for <Name>' -TestCases $testCases {
+        $functionFile = Get-ChildItem -Path $sourcePath -Recurse -Include "$Name.ps1"
+
+        $scriptFileRawContent = Get-Content -Raw -Path $functionFile.FullName
+
+        $abstractSyntaxTree = [System.Management.Automation.Language.Parser]::ParseInput($scriptFileRawContent, [ref] $null, [ref] $null)
+
+        $astSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
+
+        $parsedFunction = $abstractSyntaxTree.FindAll( $astSearchDelegate, $true ) |
+            Where-Object -FilterScript {
+                $_.Name -eq $Name
+            }
+
+        $functionHelp = $parsedFunction.GetHelpContent()
+
+        $functionHelp.Description.Length | Should -BeGreaterThan 40
+    }
+
+    It 'Should have at least one (1) example for <Name>' -TestCases $testCases {
+        $functionFile = Get-ChildItem -Path $sourcePath -Recurse -Include "$Name.ps1"
+
+        $scriptFileRawContent = Get-Content -Raw -Path $functionFile.FullName
+
+        $abstractSyntaxTree = [System.Management.Automation.Language.Parser]::ParseInput($scriptFileRawContent, [ref] $null, [ref] $null)
+
+        $astSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
+
+        $parsedFunction = $abstractSyntaxTree.FindAll( $astSearchDelegate, $true ) |
+            Where-Object -FilterScript {
+                $_.Name -eq $Name
+            }
+
+        $functionHelp = $parsedFunction.GetHelpContent()
+
+        $functionHelp.Examples.Count | Should -BeGreaterThan 0
+        $functionHelp.Examples[0] | Should -Match ([regex]::Escape($function.Name))
+        $functionHelp.Examples[0].Length | Should -BeGreaterThan ($function.Name.Length + 10)
+
+    }
+
+    It 'Should have described all parameters for <Name>' -TestCases $testCases {
+        $functionFile = Get-ChildItem -Path $sourcePath -Recurse -Include "$Name.ps1"
+
+        $scriptFileRawContent = Get-Content -Raw -Path $functionFile.FullName
+
+        $abstractSyntaxTree = [System.Management.Automation.Language.Parser]::ParseInput($scriptFileRawContent, [ref] $null, [ref] $null)
+
+        $astSearchDelegate = { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }
+
+        $parsedFunction = $abstractSyntaxTree.FindAll( $astSearchDelegate, $true ) |
+            Where-Object -FilterScript {
+                $_.Name -eq $Name
+            }
+
+        $functionHelp = $parsedFunction.GetHelpContent()
+
+        $parameters = $parsedFunction.Body.ParamBlock.Parameters.Name.VariablePath.ForEach({ $_.ToString() })
+
+        foreach ($parameter in $parameters)
+        {
+            $functionHelp.Parameters.($parameter.ToUpper()) | Should -Not -BeNullOrEmpty -Because ('the parameter {0} must have a description' -f $parameter)
+            $functionHelp.Parameters.($parameter.ToUpper()).Length | Should -BeGreaterThan 25 -Because ('the parameter {0} must have descriptive description' -f $parameter)
+        }
+    }
+}